Add syntax-local-binding, a generalized version of syntax-local-value
First-class definition contexts provide general-purpose control of the
local binding context, but syntax-local-value only provides a limited
way to inspect it: it cannot distinguish between out-of-context
identifiers and in-context variables. This commit adds
syntax-local-binding, which provides precisely the missing
functionality.

This commit also adds an #:immediate? keyword argument to
syntax-local-value to make it more consistent with syntax-local-binding
and eliminating the need for a separate syntax-local-value/immediate
function.