don't duplicate language as a `require`
Commit 0d17fc1b4c was meant to detangle access to requires from the
way that they're injected into the module, but it looks like it
accidentally preserved some tangling so that the language module could
be duplicated as a `#%require` in the module body. That disables
shadowing of bindings, which matters, for example, with the teaching
languages and shadowing `image?` via `2htdp/image`.