core: enhance `raise-argument-error`
A new `raise-argument-error` procedure is defined with `#:more-info`,
a new optional keyword argument to add more detail about the bad
argument's role to the `exn:fail:contract` exception constructed
by `raise-argument-error`.

The new procedure is exported by `racket/base` and shadows the
old primitive `raise-argument-error` which does not have keyword
support.

Also included in the commit is a simple error reporting infrastructre
used to assemble an error message following Racket's documented
conventions. The infrastructure is used to construct appropriate
messages for the two different ways `raise-argument-error` can be used
and also to report misuses of the new keyword argument in a format
that display keyword arguments in a clean manner.

The reference documentation for `raise-argument-error` is updated
with an explanation of the new `#:more-info` argument and an example
is included. A history annotation is also added to document the change.

New test section labelled "error" is added to `racket-test-core` to
exercise the new `raise-argument-error` to ensure backward compatibility
with existing behavior is preserved and also to verify the error
reporting infrastructure works properly.